workers/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bandit==1.8.5
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
fastapi==0.115.14
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
packageurl-python==0.17.1
pbr==6.1.1
pika==1.3.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
setuptools==80.9.0
sniffio==1.3.1
starlette==0.46.2
stevedore==5.4.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
